% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esri_helpers.R
\name{etl_list_gdb_layers}
\alias{etl_list_gdb_layers}
\title{List layers in a File Geodatabase}
\usage{
etl_list_gdb_layers(gdb_path)
}
\arguments{
\item{gdb_path}{Path to .gdb folder.}
}
\value{
A tibble of layers.
}
\description{
List layers in a File Geodatabase
}
