% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esri_io.R
\name{guess_spatial_format}
\alias{guess_spatial_format}
\title{Guess spatial file format}
\usage{
guess_spatial_format(dsn)
}
\arguments{
\item{dsn}{Path to a spatial data source.}
}
\value{
One of \code{"gdb"}, \code{"shapefile"}, or \code{"gpkg"}.
}
\description{
Guess spatial file format
}
\keyword{internal}
