% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_spatial.R
\name{qa_spatial_plot}
\alias{qa_spatial_plot}
\title{Quick spatial QA plot}
\usage{
qa_spatial_plot(x, quiet = FALSE)
}
\arguments{
\item{x}{An sf object.}

\item{quiet}{Logical. If \code{TRUE}, suppress messages.}
}
\value{
A ggplot object.
}
\description{
Produces a simple plot of spatial features for quick visual inspection.
}
