% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qs2_io.R
\name{read_sf_from_qs2}
\alias{read_sf_from_qs2}
\title{Read sf object from qs2}
\usage{
read_sf_from_qs2(path, quiet = FALSE)
}
\arguments{
\item{path}{Input qs2 file path.}

\item{quiet}{Logical. If \code{TRUE}, suppress status messages.}
}
\value{
An sf object.
}
\description{
Read sf object from qs2
}
